[Tags] Move data classes and helper functions out
- Move TagAlias, TagEncoder, and TagInfo to data.py.
- Move tag_decoder to helpers.py.
